added markers, favourites and delete map

+++ db/schema/02_maps.sql
@@ -3,7 +3,7 @@ CREATE TABLE maps (
   id SERIAL PRIMARY KEY NOT NULL,
   title VARCHAR(255) NOT NULL,
   user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-  latitude FLOAT NOT NULL,
-  longitude FLOAT NOT NULL,
+  latitude DECIMAL NOT NULL,
+  longitude DECIMAL NOT NULL,
   active_map BOOLEAN DEFAULT TRUE
 );
